Move fakes to respective packages.
