+++ coil-fakeimageloader/src/main/java/coil/util/Utils.kt
